% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_split_violin.R
\name{geom_split_violin}
\alias{geom_split_violin}
\title{Split violin plot}
\usage{
geom_split_violin(mapping = NULL, data = NULL, stat = "ydensity",
  position = "identity", ..., draw_quantiles = NULL, trim = TRUE,
  scale = "area", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)
}
\description{
TBA, but it is very similar to geom_violin()
   Taken from https://gist.github.com/Karel-Kroeze/746685f5613e01ba820a31e57f87ec87
}
\examples{
ggplot(mtcars, aes(x = wt, y = mpg, colour = factor(gear))) +
geom_split_violin()
}
