% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_split_violin.R
\docType{data}
\name{GeomSplitViolin}
\alias{GeomSplitViolin}
\title{Utility function that is used by geom_split_violin()}
\format{An object of class \code{GeomSplitViolin} (inherits from \code{GeomViolin}, \code{Geom}, \code{ggproto}, \code{gg}) of length 2.}
\usage{
GeomSplitViolin
}
\description{
TBA, taken from https://gist.github.com/Karel-Kroeze/746685f5613e01ba820a31e57f87ec87
}
\keyword{datasets}
