% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_janco.R
\name{theme_janco_point}
\alias{theme_janco_point}
\title{Minimal theme for scatterplots (etc)}
\usage{
theme_janco_point(base_size = 12, base_family = "sans")
}
\arguments{
\item{base_size}{global font size (number)}

\item{base_family}{global font family (e.g. sans, serif)}
}
\value{
Adds a theme to a ggplot
}
\description{
A minimal theme based on ggthemes::theme_fivethirtyeight()
    This theme has both horizontal and vertical panel.grids
}
\examples{
ggplot(mtcars, aes(x = wt, y = mpg, colour = factor(gear))) +
geom_point() +
theme_janco_point()
}
